Seleccionar una cita

+++ src/AppBundle/Controller/CitaController.php
@@ -118,6 +118,12 @@ class CitaController extends Controller
         }
 
         $cita = $repository->findByAlumnoCita($alumno, $id);
+
+        if($cita->getAsesor() != null){
+            $this->addFlash('danger','Está cita no se puede eliminar.');
+            return $this->redirectToRoute('citas');
+        }
+
         if($cita){
             $em = $this->getDoctrine()->getManager();
             $em->remove($cita);
@@ -164,6 +170,12 @@ class CitaController extends Controller
      * @Route("/solicitud/detalles/{id}", name="solicitudes_detalles")
      */
     public function detallesAction(Request $request, $id){
+
+        if(!$this->getUser()->isComplete()){
+            $this->addFlash('danger','No puedes aceptar una cita si no has completado tu perfil.');
+            return $this->redirectToRoute('registro');
+        }
+
         $repository = $this->getDoctrine()->getRepository(Cita::class);
         $cita = $repository->findOneById($id);
